refactor(toolkit)!: split connector toolkit types
split connector toolkit types
